backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plutus"
version = "0.1.0"
description = "Production-grade reconciliation and reporting platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Plutus Team" }
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",

    # Validation & Serialization
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",

    # Background Tasks
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",

    # Data Processing
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "openpyxl>=3.1.2",
    "xlrd>=2.0.1",

    # Matching & ML
    "rapidfuzz>=3.6.1",
    "scikit-learn>=1.4.0",

    # Export
    "reportlab>=4.0.8",
    "weasyprint>=60.2",

    # AI Providers
    "httpx>=0.26.0",
    "anthropic>=0.18.0",
    "openai>=1.10.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "aiosqlite>=0.20.0",

    # Linting & Formatting
    "ruff>=0.1.14",
    "black>=24.1.0",
    "mypy>=1.8.0",

    # Type stubs
    "types-python-dateutil>=2.8.19",
    "types-passlib>=1.7.7",
    "pandas-stubs>=2.1.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "rapidfuzz.*",
    "reportlab.*",
    "weasyprint.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
